refactor: new component Cart + refactor TableForSale
