+++ README.md
@@ -23,8 +23,9 @@
 11. ЛИШНИЕ ОБЕРТКИ В КОДЕ УДАЛЕНЫ
 12. Вынесла константы в отдельный файл, переименовала некотрые переменные чтобы интуитивно стало понятно где компонент
 13. TableToSale.js должен соответствовать свой файл TableToSale.css, а не какой-то Tables.css. Переименовывала
-14. C корзиной: переименовала один хук, чтобы было понятно что он на открытие корзины
+14. C корзиной: переименовала один хук, чтобы было понятно что он на открытие корзины. Подняла стейдж с пропсами, прокинула их в хедер и в НОВЫЙ компонент Cart.
 
 **Для себя:**
 
 1.  MainDrawer - подумать как вынесли константы и стоит ли
+2.  НЕ ЗАБУДЬ ПЕРЕИМЕНОВАТЬ В ТАБЛИЦАХ (ИЛИ В КОРЗИНЕ, ЧЕКНИ СТИЛИ ПО БЭМ)
